workshops/RAG_workshop/pyproject.toml
[project]
name = "rag-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.3.4",
    "fastmcp>=2.13.0.2",
    "gensim>=4.4.0",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "mcp>=1.0.0",
    "nltk>=3.9.2",
    "numpy>=2.3.4",
    "openai>=2.7.1",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.2",
    "snowflake-connector-python>=3.0.0",
    "tabulate>=0.9.0",
    "torch>=2.9.0",
    "transformers>=4.57.1",
]
